Notification Feature

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -58,7 +58,19 @@ export async function POST(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    await prisma.like.upsert({
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select:{
+        userId: true
+      }
+    })
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+    
+    await prisma.$transaction([
+    prisma.like.upsert({
       where: {
         userId_postId: {
           userId: loggedInUser.id,
@@ -70,7 +82,21 @@ export async function POST(
         postId,
       },
       update: {},
-    });
+    }),
+    ...(loggedInUser.id !== post.userId
+      ? [
+        prisma.notification.create({
+          data:{
+            issuerId: loggedInUser.id,
+            recipientId: post.userId,
+            postId,
+            type: "LIKE"
+          }
+        })
+      ] :
+      []
+    )
+  ]);
 
     return new Response();
   } catch (error) {
@@ -90,12 +116,33 @@ export async function DELETE(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    await prisma.like.deleteMany({
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
+      },
+    });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+    prisma.like.deleteMany({
       where: {
         userId: loggedInUser.id,
         postId,
       },
-    });
+    }),
+    prisma.notification.deleteMany({
+      where: {
+        issuerId: loggedInUser.id,
+        recipientId: post.userId,
+        postId,
+        type: "LIKE",
+      },
+    }),
+  ])
 
     return new Response();
   } catch (error) {
